requirements.txt
python-telegram-bot==22.3
requests==2.31.0
python-dotenv==1.0.0
setuptools
pytz
APScheduler==3.10.4 
Pillow>=10.2.0
pytesseract==0.3.13
qrcode[pil]
